requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-gis==1.0
GDAL @ file:///C:/xampp/htdocs/django-geoserver/GDAL-3.4.3-cp311-cp311-win_amd64.whl
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
